Fix type annotations and platform modules for CI/CD
- Add List to typing imports in loader.py
- Add Tuple to typing imports in safety.py
- Fix Config() calls to use Config.load()
- Create platform module placeholders (windows.py, linux.py, macos.py)
- Fix mypy.ini configuration (remove duplicate sections)
- Fix Windows path strings to use raw strings
- Fix return type annotations
- Create examples/ directory

Fixes for GitHub Actions CI/CD compatibility.

+++ diskcleaner/core/cache.py
@@ -202,5 +202,5 @@ class CacheManager:
 
         return {
             "total_files": len(cache_files),
-            "total_size_mb": round(total_size / (1024 * 1024), 2),
+            "total_size_mb": int(round(total_size / (1024 * 1024), 2)),
         }
